Learn interest weights from clicks and score auctions with them
Adds user_interests.weight and makes score_campaign sum a user's real
per-interest weights instead of counting matched interests, so an
interest someone actually clicks on outranks one they merely happen to
share with a campaign. Eligibility is deliberately untouched - weight
shifts the odds among already-eligible campaigns, it never decides
whether a campaign is eligible at all.

The weights are fetched once per request and the same dict is passed to
both pick_winner and record_win rather than re-queried, so the score
persisted on the impression is always the score that actually chose the
winner.

Impression.score becomes a Float: summing weights that decay produces
values an Integer column cannot represent.

Also fixes two real problems in the same write path:

- campaign.spent += COST_PER_WIN was a Python-side read-modify-write, so
  two workers winning the same campaign concurrently both read the same
  value and one charge vanished. A forced-collision test lost 45 of 50
  charges. Now an atomic UPDATE ... SET spent = spent + ... so Postgres
  does the arithmetic under a row lock.
- record_win scored a re-fetched ORM Campaign, lazy-loading targeting and
  its interests when the caller already had both in memory, and then
  refreshed the impression it had just written. Scoring the in-memory
  candidate and setting expire_on_commit=False removes three queries per
  win (7 statements down to 3).

+++ app/models/impression.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Table, Uuid, Enum as SqlEnum
+from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String, Table, Uuid, Enum as SqlEnum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -23,6 +23,16 @@ class Impression(Base):
 
     __tablename__ = "impressions"
     id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
+    # Deliberately NOT indexed, after measuring both sides - see §30.3.
+    # §28 added indexes here expecting the analytics queries to need them;
+    # measurement showed the dashboard's GROUP BY correctly ignores them
+    # (it reads every row anyway), while they cost ~0.60ms of index
+    # maintenance on EVERY impression insert - the single hottest write in
+    # the system. Their only real beneficiaries would be per-campaign
+    # drill-down endpoints that don't exist and user/campaign deletes that
+    # never happen (seed.py uses TRUNCATE, which bypasses FK checks), so
+    # they were dropped again in migration `dcf2afa14820`. Re-add them the day a
+    # query actually needs them, not before.
     user_id: Mapped[uuid.UUID] = mapped_column(Uuid, ForeignKey("users.id", ondelete="RESTRICT"), nullable=False)
     campaign_id: Mapped[uuid.UUID] = mapped_column(Uuid, ForeignKey("campaigns.id", ondelete="RESTRICT"), nullable=False)
     occurred_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
@@ -32,5 +42,9 @@ class Impression(Base):
         SqlEnum(DeviceType, values_callable=lambda enum_cls: [e.value for e in enum_cls]),
         nullable=False,
     )
-    score: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Float, not Integer: since Step 8b (STUDY_NOTES.md §22), score_campaign
+    # sums real per-interest weights (nudged by clicks, not always whole
+    # numbers once decay - see scripts/consume_events.py - has been
+    # applied) rather than counting whole matched interests.
+    score: Mapped[float] = mapped_column(Float, nullable=False)
     interests: Mapped[list[Interest]] = relationship(secondary=impression_interests)
